# Merging: Treat inlines as if they were static when mergeInlines is off