This should hopefully fix all issues with merging of inlines. The logic now is as follows:

If `merge_inlines` is off (default), inline functions are handled as if they were static, meaning they will always be renamed. This should fix any unintended instances of multiple `fundec`s for one given `varinfo`.