@michael-schwarz 2022-03-27
Actually, all that was needed to get better number is wrapping the merging into a call to `Stas.time` inside Goblint.