@michael-schwarz 2022-03-27
With merging of inlines:

```
  mergeCIL                       285.114 s
  analysis                        0.000 s
    createCFG                      14.989 s
      handle                          4.853 s
      iter_connect                    8.983 s
        computeSCCs                     4.394 s
```

Without it:

```
  mergeCIL                       66.626 s
  analysis                        0.000 s
    createCFG                      71.905 s
      handle                          9.731 s
      iter_connect                   57.776 s
        computeSCCs                     7.778 s
```

So it looks like it buys us 2x speedup to not merge inlines (at least until the start of analysis).